Move async stream members into own file
